pyproject.toml
[project]
name = "rowdybottypiper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.44",
    "botocore>=1.42.44",
    "bs4>=0.0.2",
    "paramiko>=4.0.0",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "selenium>=4.40.0",
    "slack-sdk>=3.39.0",
]

[project.scripts]
rbp = "rowdybottypiper.cli.main:main"
